[#203] refactor(library)!: move 'expr(...)' to 'expression' package
This is a part of preparation for introducing type-safe expression DSL.
