+++ library/src/test/kotlin/it/krzeminski/githubactions/dsl/WorkflowBuilderTest.kt
@@ -7,6 +7,7 @@ import io.kotest.matchers.shouldBe
 import it.krzeminski.githubactions.domain.RunnerType
 import it.krzeminski.githubactions.domain.RunnerType.UbuntuLatest
 import it.krzeminski.githubactions.domain.triggers.Push
+import it.krzeminski.githubactions.dsl.expressions.expr
 import it.krzeminski.githubactions.yaml.toYaml
 import java.nio.file.Paths
 
@@ -158,12 +159,12 @@ class WorkflowBuilderTest : FunSpec({
                 # This file was generated using Kotlin DSL (.github/workflows/some_workflow.main.kts).
                 # If you want to modify the workflow, please change the Kotlin file and regenerate this YAML file.
                 # Generated with https://github.com/krzema12/github-actions-kotlin-dsl
-                
+
                 name: Test workflow
-                
+
                 on:
                   push:
-                
+
                 jobs:
                   "test_job":
                     runs-on: "ubuntu-latest"
